A combination of languages to detect plegiarism
It has little explanation but more code, which has some hashtag explanations. I combine all languages so that it can be run either on Visual Studio, Google Collaboration, or even a Jupyter Notebook with little or no modification. in case of any complications or error messages check the IDE for the required instalation.